Moving event overview display classes.
