+++ app/src/main/java/com/adammcneilly/pocketleague/eventoverview/ui/EventOverviewViewState.kt
@@ -1,6 +1,7 @@
 package com.adammcneilly.pocketleague.eventoverview.ui
 
 import com.adammcneilly.pocketleague.core.ui.UIText
+import com.adammcneilly.pocketleague.shared.eventoverview.EventOverviewDisplayModel
 
 /**
  * A collection of possible view states for [EventOverviewScreen].
