+++ app/src/main/java/com/adammcneilly/pocketleague/standings/ui/StandingsList.kt
@@ -8,6 +8,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.adammcneilly.pocketleague.android.design.theme.PocketLeagueTheme
 import com.adammcneilly.pocketleague.core.ui.Material3Divider
+import com.adammcneilly.pocketleague.shared.eventoverview.StandingsDisplayModel
+import com.adammcneilly.pocketleague.shared.eventoverview.StandingsPlacementDisplayModel
 
 /**
  * Shows a collection of [standings] for some event.
