+++ app/src/main/java/com/adammcneilly/pocketleague/standings/ui/StandingsPlacementListItem.kt
@@ -24,6 +24,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.adammcneilly.pocketleague.android.design.theme.PocketLeagueTheme
 import com.adammcneilly.pocketleague.core.ui.PocketLeagueImage
+import com.adammcneilly.pocketleague.shared.eventoverview.StandingsPlacementDisplayModel
 
 private const val PLACEMENT_WEIGHT = 1F
 private const val TEAM_WEIGHT = 7F
